+++ tests/unit_tests/test_search_plan.py
@@ -10,14 +10,14 @@ from alphadia.search_plan import SearchPlan
 
 MOCK_DEFAULT_CONFIG = {
     "general": {
-        "adaptation_step_enabled": False,
+        "transfer_step_enabled": False,
         "adaptation_method": "transfer",
         "mbr_step_enabled": False,
     },
 }
 
 MOCK_MULTISTEP_CONFIG = {
-    "adaptation": {"some_adaptation_config_key": "some_adaptation_config_value"},
+    "transfer": {"some_adaptation_config_key": "some_adaptation_config_value"},
     "library": {"some_library_config_key": "some_library_config_value"},
     "mbr": {"some_mbr_config_key": "some_mbr_config_value"},
 }
@@ -104,7 +104,7 @@ def test_runs_plan_without_transfer_and_mbr_steps_none_dirs(
 
 
 @pytest.mark.parametrize(
-    ("adaptation_step_enabled", "adaptation_method", "mbr_step_enabled"),
+    ("transfer_step_enabled", "adaptation_method", "mbr_step_enabled"),
     [
         (True, "transfer", True),
         (True, "tto", False),
@@ -115,7 +115,7 @@ def test_runs_plan_without_transfer_and_mbr_steps_none_dirs(
 @patch("alphadia.search_plan.reporting.init_logging")
 def test_default_correctly_read(
     mock_init_logging,
-    adaptation_step_enabled,
+    transfer_step_enabled,
     adaptation_method,
     mbr_step_enabled,
 ):
@@ -127,7 +127,7 @@ def test_default_correctly_read(
         [
             {
                 "general": {
-                    "adaptation_step_enabled": adaptation_step_enabled,
+                    "transfer_step_enabled": transfer_step_enabled,
                     "adaptation_method": adaptation_method,
                     "mbr_step_enabled": mbr_step_enabled,
                 }
@@ -136,7 +136,7 @@ def test_default_correctly_read(
         ],
     )
 
-    assert search_plan._adaptation_step_enabled == adaptation_step_enabled
+    assert search_plan._transfer_step_enabled == transfer_step_enabled
     assert search_plan._adaptation_method == adaptation_method
     assert search_plan._mbr_step_enabled == mbr_step_enabled
 
@@ -148,11 +148,95 @@ def test_raises_on_invalid_adaptation_method(mock_init_logging):
         get_search_plan(
             {
                 "general": {
-                    "adaptation_step_enabled": True,
+                    "transfer_step_enabled": True,
                     "adaptation_method": "invalid_method",
                 }
             }
         )
+
+
+@patch("alphadia.search_plan.reporting.init_logging")
+def test_old_config_no_adaptation_method_defaults_to_transfer(mock_init_logging):
+    """Backward compat: user config has transfer_step_enabled but no adaptation_method.
+
+    adaptation_method should default to "transfer" from default.yaml so the pipeline
+    behaves identically to the pre-adaptation_method era.
+    """
+    search_plan = get_search_plan(
+        {"general": {"transfer_step_enabled": True}},
+        {},
+        [MOCK_DEFAULT_CONFIG, MOCK_MULTISTEP_CONFIG],
+    )
+
+    assert search_plan._adaptation_method == "transfer"
+    assert search_plan._transfer_step_enabled is True
+
+
+@patch("alphadia.search_plan.reporting.init_logging")
+def test_old_config_no_transfer_step_defaults_to_disabled(mock_init_logging):
+    """Backward compat: completely minimal user config with no step flags at all.
+
+    Both transfer_step_enabled and adaptation_method should fall back to defaults
+    (False and "transfer" respectively), so a single library step runs.
+    """
+    search_plan = get_search_plan({}, {}, [MOCK_DEFAULT_CONFIG, MOCK_MULTISTEP_CONFIG])
+
+    assert search_plan._transfer_step_enabled is False
+    assert search_plan._adaptation_method == "transfer"
+    assert search_plan._mbr_step_enabled is False
+
+
+@patch("alphadia.search_plan.reporting.init_logging")
+@patch("alphadia.search_plan.SearchStep")
+@patch("alphadia.search_plan.SearchPlan._get_optimized_values_config")
+def test_old_config_runs_transfer_path_without_adaptation_method(
+    mock_get_dyn_config, mock_plan, mock_init_logging
+):
+    """Backward compat: user config with transfer_step_enabled but no adaptation_method
+    should run exactly the same as explicitly setting adaptation_method: "transfer".
+    """
+    # Old-style config: no adaptation_method key at all
+    old_user_config = {"general": {"transfer_step_enabled": True}}
+
+    search_plan = get_search_plan(old_user_config)
+
+    dynamic_config = {"some_dynamic_config_key": "some_dynamic_config_value"}
+    mock_get_dyn_config.return_value = dynamic_config
+
+    search_plan.run_plan()
+
+    # transfer step should use transfer_learning (the pre-TTO default path)
+    assert mock_plan.call_args_list[0].kwargs == {
+        "output_folder": _convert_path("/user_provided_output_path/transfer"),
+        "config": old_user_config,
+        "extra_config": MOCK_MULTISTEP_CONFIG["transfer"]
+        | {"transfer_learning": {"enabled": True}},
+        "cli_config": BASE_CLI_PARAMS_CONFIG,
+        "step_name": "transfer",
+    }
+
+    # library step should receive the peptdeep model path (transfer method output)
+    assert mock_plan.call_args_list[1].kwargs == {
+        "output_folder": _convert_path("/user_provided_output_path"),
+        "config": old_user_config,
+        "extra_config": MOCK_MULTISTEP_CONFIG["library"]
+        | {
+            "library_prediction": {
+                "peptdeep_model_path": _convert_path(
+                    "/user_provided_output_path/transfer/peptdeep.transfer"
+                ),
+                "enabled": True,
+                "use_peptdeep_kontext": False,
+            },
+        }
+        | dynamic_config,
+        "cli_config": BASE_CLI_PARAMS_CONFIG,
+        "step_name": "library",
+    }
+
+    mock_get_dyn_config.assert_called_once_with(
+        Path("/user_provided_output_path/transfer")
+    )
 
 
 @patch("alphadia.search_plan.reporting.init_logging")
@@ -164,7 +248,7 @@ def test_runs_plan_with_adaptation_step_transfer(
     """Test that the SearchPlan object runs the plan correctly with the adaptation step (transfer method)."""
     additional_user_config = {
         "general": {
-            "adaptation_step_enabled": True,
+            "transfer_step_enabled": True,
             "adaptation_method": "transfer",
             "mbr_step_enabled": False,
         }
@@ -182,12 +266,12 @@ def test_runs_plan_with_adaptation_step_transfer(
 
     # adaptation_step
     assert mock_plan.call_args_list[0].kwargs == {
-        "output_folder": _convert_path("/user_provided_output_path/adaptation"),
+        "output_folder": _convert_path("/user_provided_output_path/transfer"),
         "config": BASE_USER_CONFIG | additional_user_config,
-        "extra_config": MOCK_MULTISTEP_CONFIG["adaptation"]
+        "extra_config": MOCK_MULTISTEP_CONFIG["transfer"]
         | {"transfer_learning": {"enabled": True}},
         "cli_config": BASE_CLI_PARAMS_CONFIG,
-        "step_name": "adaptation",
+        "step_name": "transfer",
     }
 
     # library_step
@@ -198,7 +282,7 @@ def test_runs_plan_with_adaptation_step_transfer(
         | {
             "library_prediction": {
                 "peptdeep_model_path": _convert_path(
-                    "/user_provided_output_path/adaptation/peptdeep.transfer"
+                    "/user_provided_output_path/transfer/peptdeep.transfer"
                 ),
                 "enabled": True,
                 "use_peptdeep_kontext": False,
@@ -211,7 +295,7 @@ def test_runs_plan_with_adaptation_step_transfer(
 
     assert mock_plan.return_value.run.call_count == 2
     mock_get_dyn_config.assert_called_once_with(
-        Path("/user_provided_output_path/adaptation")
+        Path("/user_provided_output_path/transfer")
     )
 
 
@@ -224,7 +308,7 @@ def test_runs_plan_with_adaptation_step_tto(
     """Test that the SearchPlan object runs the plan correctly with the adaptation step (tto method)."""
     additional_user_config = {
         "general": {
-            "adaptation_step_enabled": True,
+            "transfer_step_enabled": True,
             "adaptation_method": "tto",
             "mbr_step_enabled": False,
         }
@@ -242,12 +326,12 @@ def test_runs_plan_with_adaptation_step_tto(
 
     # adaptation_step
     assert mock_plan.call_args_list[0].kwargs == {
-        "output_folder": _convert_path("/user_provided_output_path/adaptation"),
+        "output_folder": _convert_path("/user_provided_output_path/transfer"),
         "config": BASE_USER_CONFIG | additional_user_config,
-        "extra_config": MOCK_MULTISTEP_CONFIG["adaptation"]
+        "extra_config": MOCK_MULTISTEP_CONFIG["transfer"]
         | {"context_extraction": {"enabled": True}},
         "cli_config": BASE_CLI_PARAMS_CONFIG,
-        "step_name": "adaptation",
+        "step_name": "transfer",
     }
 
     # library_step
@@ -258,7 +342,7 @@ def test_runs_plan_with_adaptation_step_tto(
         | {
             "library_prediction": {
                 "context_path": _convert_path(
-                    "/user_provided_output_path/adaptation/peptdeep_kontext.context"
+                    "/user_provided_output_path/transfer/peptdeep_kontext.context"
                 ),
                 "enabled": True,
                 "use_peptdeep_kontext": True,
@@ -271,7 +355,7 @@ def test_runs_plan_with_adaptation_step_tto(
 
     assert mock_plan.return_value.run.call_count == 2
     mock_get_dyn_config.assert_called_once_with(
-        Path("/user_provided_output_path/adaptation")
+        Path("/user_provided_output_path/transfer")
     )
 
 
@@ -282,7 +366,7 @@ def test_runs_plan_with_mbr_step(mock_get_dyn_config, mock_plan, mock_init_loggi
     """Test that the SearchPlan object runs the plan correctly with the mbr step enabled."""
     additional_user_config = {
         "general": {
-            "adaptation_step_enabled": False,
+            "transfer_step_enabled": False,
             "mbr_step_enabled": True,
         }
     }
@@ -336,7 +420,7 @@ def test_runs_plan_with_adaptation_transfer_and_mbr_steps(
     """Test that the SearchPlan object runs the plan correctly with both the adaptation (transfer) and mbr steps enabled."""
     additional_user_config = {
         "general": {
-            "adaptation_step_enabled": True,
+            "transfer_step_enabled": True,
             "adaptation_method": "transfer",
             "mbr_step_enabled": True,
         }
@@ -354,12 +438,12 @@ def test_runs_plan_with_adaptation_transfer_and_mbr_steps(
 
     # adaptation_step
     assert mock_plan.call_args_list[0].kwargs == {
-        "output_folder": _convert_path("/user_provided_output_path/adaptation"),
+        "output_folder": _convert_path("/user_provided_output_path/transfer"),
         "config": BASE_USER_CONFIG | additional_user_config,
-        "extra_config": MOCK_MULTISTEP_CONFIG["adaptation"]
+        "extra_config": MOCK_MULTISTEP_CONFIG["transfer"]
         | {"transfer_learning": {"enabled": True}},
         "cli_config": BASE_CLI_PARAMS_CONFIG,
-        "step_name": "adaptation",
+        "step_name": "transfer",
     }
 
     # library_step
@@ -370,7 +454,7 @@ def test_runs_plan_with_adaptation_transfer_and_mbr_steps(
         | {
             "library_prediction": {
                 "peptdeep_model_path": str(
-                    Path("/user_provided_output_path/adaptation/peptdeep.transfer")
+                    Path("/user_provided_output_path/transfer/peptdeep.transfer")
                 ),
                 "enabled": True,
                 "use_peptdeep_kontext": False,
@@ -398,7 +482,7 @@ def test_runs_plan_with_adaptation_transfer_and_mbr_steps(
 
     assert mock_plan.return_value.run.call_count == 3
     mock_get_dyn_config.assert_called_once_with(
-        Path("/user_provided_output_path/adaptation")
+        Path("/user_provided_output_path/transfer")
     )
 
 
@@ -411,7 +495,7 @@ def test_runs_plan_with_adaptation_tto_and_mbr_steps(
     """Test that the SearchPlan object runs the plan correctly with both the adaptation (tto) and mbr steps enabled."""
     additional_user_config = {
         "general": {
-            "adaptation_step_enabled": True,
+            "transfer_step_enabled": True,
             "adaptation_method": "tto",
             "mbr_step_enabled": True,
         }
@@ -429,12 +513,12 @@ def test_runs_plan_with_adaptation_tto_and_mbr_steps(
 
     # adaptation_step
     assert mock_plan.call_args_list[0].kwargs == {
-        "output_folder": _convert_path("/user_provided_output_path/adaptation"),
+        "output_folder": _convert_path("/user_provided_output_path/transfer"),
         "config": BASE_USER_CONFIG | additional_user_config,
-        "extra_config": MOCK_MULTISTEP_CONFIG["adaptation"]
+        "extra_config": MOCK_MULTISTEP_CONFIG["transfer"]
         | {"context_extraction": {"enabled": True}},
         "cli_config": BASE_CLI_PARAMS_CONFIG,
-        "step_name": "adaptation",
+        "step_name": "transfer",
     }
 
     # library_step
@@ -445,7 +529,7 @@ def test_runs_plan_with_adaptation_tto_and_mbr_steps(
         | {
             "library_prediction": {
                 "context_path": str(
-                    Path("/user_provided_output_path/adaptation/peptdeep_kontext.context")
+                    Path("/user_provided_output_path/transfer/peptdeep_kontext.context")
                 ),
                 "enabled": True,
                 "use_peptdeep_kontext": True,
@@ -473,7 +557,7 @@ def test_runs_plan_with_adaptation_tto_and_mbr_steps(
 
     assert mock_plan.return_value.run.call_count == 3
     mock_get_dyn_config.assert_called_once_with(
-        Path("/user_provided_output_path/adaptation")
+        Path("/user_provided_output_path/transfer")
     )
